PC port: Nintendo logo rendering + 64-bit fixes + GX rendering pipeline
Major milestone: first visible rendered output on PC (Nintendo logo on black background).

Rendering fixes:
- Fix operator delete to check JKR arena range before routing to JKRHeap::free,
  preventing crashes when Metal/system frameworks free their own memory
- Fix ~30 files with 64-bit pointer truncation (u32/s32 casts → uintptr_t)
- Fix OSCachedToPhysical/OSRoundUp32B macros to identity-map on PC
- Fix GXColor byte order: add GXCOLOR_R/G/B/A macros for LE struct-to-u32
  extraction in GXSetTevColor, GXSetCopyClear, GXSetChanMatColor, etc.
- Fix texcoord fixed-point: GXSetVtxAttrFmt stores frac bits, GXTexCoord2s16
  applies the divisor (was passing raw integers as float coords)
- Fix shader loading: search multiple paths (shaders/, bin/shaders/)
- Fix uniform caching: call pc_gx_cache_uniform_locations at init + dirty=ALL
- Fix default GL cull mode: glDisable(GL_CULL_FACE) matching GX default
- Force PC_GX_DIRTY_ALL on every flush to ensure uniform upload
- Fix clear color: use game's mBackColor instead of hardcoded debug color
- Pre-warm 5 Metal shader pipeline variants at init to reduce SIGBUS
- Improve signal handler to detect Metal/system framework crashes via dladdr
- Clean up per-frame debug logging (gate behind g_pc_verbose)
- Remove debug prints from J3DSys::reinitGX and J3DSys::drawInit

YAZ0 decompression fix:
- Fix buffer reallocation when expandSize was capped by dstLength but the
  actual Yaz0 header specifies a larger size (prevents buffer overflow)

Other fixes:
- Fix mDoDvdThd::DVDLogoMode defined unconditionally (was #if DEBUG only)
- Add g_printOtherHeapDebug definition
- Fix d_a_npc_henna.cpp, d_a_npc_tkj2.cpp, d_bg_w_kcol.cpp, d_resorce.cpp,
  m_Do_ext.cpp pointer casts
- Delete stale pc/build directory

+++ libs/JSystem/src/JAudio2/JASAramStream.cpp
@@ -583,7 +583,7 @@ void JASAramStream::updateChannel(u32 i_callbackType, JASChannel* i_channel,
 s32 JASAramStream::channelProc() {
     OSMessage msg;
     while (OSReceiveMessage(&field_0x020, &msg, OS_MESSAGE_NOBLOCK)) {
-        switch ((u32)msg) {
+        switch ((uintptr_t)msg) {
         case 4:
             field_0x0ac = true;
             break;
@@ -592,18 +592,18 @@ s32 JASAramStream::channelProc() {
             break;
         }
     }
-    
+
     if (!field_0x0ac) {
         return 0;
     }
 
     while (OSReceiveMessage(&field_0x000, &msg, OS_MESSAGE_NOBLOCK)) {
-        switch ((u32)msg & 0xff) {
+        switch ((uintptr_t)msg & 0xff) {
         case 0:
             channelStart();
             break;
         case 1:
-            channelStop(JSUHiHalf((u32)msg));
+            channelStop(JSUHiHalf((u32)(uintptr_t)msg));
             break;
         case 2:
             field_0x0ae |= 1;
