+++ libs/JSystem/src/JUtility/JUTCacheFont.cpp
@@ -351,7 +351,7 @@ void JUTCacheFont::getGlyphFromAram(JUTCacheFont::TGlyphCacheInfo* param_0,
     *param_3 = iVar2;
     *r30 -= iVar2 * iVar3;
     u8* result =
-        JKRAramToMainRam((u32)param_0->mPrev + pGylphCacheInfo->field_0x10 * iVar2, pCachePage->mImage,
+        JKRAramToMainRam((uintptr_t)param_0->mPrev + pGylphCacheInfo->field_0x10 * iVar2, pCachePage->mImage,
                          pGylphCacheInfo->field_0x10, EXPAND_SWITCH_UNKNOWN0, 0, NULL, 0xffffffff, NULL);
     JUT_ASSERT(624, result);
     GXInitTexObj(&pCachePage->mTexObj, pCachePage->mImage, pGylphCacheInfo->mWidth, pGylphCacheInfo->mHeight,
